+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleBigQueryDAO.scala
@@ -1,12 +1,14 @@
 package org.broadinstitute.dsde.workbench.google
 
-import com.google.api.services.bigquery.model.{GetQueryResultsResponse, Job, JobReference}
+import com.google.api.services.bigquery.model.{GetQueryResultsResponse, Job, JobReference, QueryParameter}
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 import scala.concurrent.Future
 
 trait GoogleBigQueryDAO {
   def startQuery(project: GoogleProject, querySql: String): Future[JobReference]
+
+  def startParameterizedQuery(project: GoogleProject, querySql: String, queryParameters: List[QueryParameter], parameterMode: String): Future[JobReference]
 
   def getQueryStatus(jobRef: JobReference): Future[Job]